✅added validations in signup page
